+++ sample_consensus/include/pcl/sample_consensus/sac_model_stick.h
@@ -177,7 +177,7 @@ namespace pcl
       {
         if (model_coefficients.size () != 7)
         {
-          PCL_ERROR ("[pcl::SampleConsensusModelStick::selectWithinDistance] Invalid number of model coefficients given (%lu)!\n", (unsigned long)model_coefficients.size ());
+          PCL_ERROR ("[pcl::SampleConsensusModelStick::selectWithinDistance] Invalid number of model coefficients given (%zu)!\n", model_coefficients.size ());
           return (false);
         }
